+++ lithops/serverless/backends/ibm_cf/ibm_cf.py
@@ -16,7 +16,6 @@
 
 import os
 import logging
-import shutil
 from threading import Lock
 
 from lithops import utils
@@ -290,7 +289,7 @@ class IBMCloudFunctionsBackend:
                 for action in actions:
                     version, image_name, memory = self._unformat_function_name(action['name'])
                     if docker_image_name == image_name or docker_image_name == 'all':
-                        runtimes.append((image_name, memory, version))
+                        runtimes.append((image_name, memory, version, action['name']))
         return runtimes
 
     def pre_invoke(self, docker_image_name, runtime_memory):